requirements.txt
# Add your Python dependencies here
# Example:
# requests>=2.25.1
# numpy>=1.21.0
# pandas>=1.3.0
langgraph==0.0.28
aiohttp==3.9.0
python-dotenv==1.0.0